+++ example/lib/data/example_lazy_stack_data.dart
@@ -1,23 +1,23 @@
 import 'package:recursive_tree_flutter/recursive_tree_flutter.dart';
 
-import '../models/custom_node_type.dart';
+import '../models/ex.dart';
 
-TreeType<CustomNodeType> createRoot() => TreeType<CustomNodeType>(
-      data: CustomNodeType.sampleInner("0"),
+TreeType<EasyNodeType> createRoot() => TreeType<EasyNodeType>(
+      data: Ex.sampleInner("0"),
       children: [],
       parent: null,
     );
 
 //? ------------------ lv 1
 
-List<TreeType<CustomNodeType>> createChildrenOfRoot() => [
-      TreeType<CustomNodeType>(
-        data: CustomNodeType.sampleInner("1.1"),
+List<TreeType<EasyNodeType>> createChildrenOfRoot() => [
+      TreeType<EasyNodeType>(
+        data: Ex.sampleInner("1.1"),
         children: [],
         parent: null,
       ),
-      TreeType<CustomNodeType>(
-        data: CustomNodeType.sampleInner("1.2"),
+      TreeType<EasyNodeType>(
+        data: Ex.sampleInner("1.2"),
         children: [],
         parent: null,
       ),
@@ -25,19 +25,19 @@ List<TreeType<CustomNodeType>> createChildrenOfRoot() => [
 
 //? ------------------ lv 2
 
-List<TreeType<CustomNodeType>> createChildrenOfLv1_1() => [
-      TreeType<CustomNodeType>(
-        data: CustomNodeType.sampleInner("2.1"),
+List<TreeType<EasyNodeType>> createChildrenOfLv1_1() => [
+      TreeType<EasyNodeType>(
+        data: Ex.sampleInner("2.1"),
         children: [],
         parent: null,
       ),
-      TreeType<CustomNodeType>(
-        data: CustomNodeType.sampleInner("2.2"),
+      TreeType<EasyNodeType>(
+        data: Ex.sampleInner("2.2"),
         children: [],
         parent: null,
       ),
-      TreeType<CustomNodeType>(
-        data: CustomNodeType.sampleInner("2.3"),
+      TreeType<EasyNodeType>(
+        data: Ex.sampleInner("2.3"),
         children: [],
         parent: null,
       ),
@@ -45,27 +45,27 @@ List<TreeType<CustomNodeType>> createChildrenOfLv1_1() => [
 
 //? ------------------ lv 3
 
-List<TreeType<CustomNodeType>> createChildrenOfLv2_1() => [
-      TreeType<CustomNodeType>(
-        data: CustomNodeType.sampleLeaf("3.1"),
+List<TreeType<EasyNodeType>> createChildrenOfLv2_1() => [
+      TreeType<EasyNodeType>(
+        data: Ex.sampleLeaf("3.1"),
         children: [],
         parent: null,
       ),
-      TreeType<CustomNodeType>(
-        data: CustomNodeType.sampleLeaf("3.2"),
+      TreeType<EasyNodeType>(
+        data: Ex.sampleLeaf("3.2"),
         children: [],
         parent: null,
       ),
-      TreeType<CustomNodeType>(
-        data: CustomNodeType.sampleLeaf("3.3"),
+      TreeType<EasyNodeType>(
+        data: Ex.sampleLeaf("3.3"),
         children: [],
         parent: null,
       ),
     ];
 
-List<TreeType<CustomNodeType>> createChildrenOfLv2_2() => [
-      TreeType<CustomNodeType>(
-        data: CustomNodeType.sampleLeaf("3.4"),
+List<TreeType<EasyNodeType>> createChildrenOfLv2_2() => [
+      TreeType<EasyNodeType>(
+        data: Ex.sampleLeaf("3.4"),
         children: [],
         parent: null,
       ),