+++ example/lib/helpers/ex_traversal_func.dart
@@ -1,0 +1,58 @@
+import 'package:flutter/material.dart';
+import 'package:recursive_tree_flutter/recursive_tree_flutter.dart';
+
+void chosenLeaves(BuildContext context, TreeType tree) {
+  List<TreeType<EasyNodeType>> result = [];
+  var root = findRoot(tree);
+  returnChosenLeaves(root, result);
+  String resultTxt = "";
+  for (var e in result) {
+    resultTxt += "${e.data.title}\n\n";
+  }
+  if (resultTxt.isEmpty) resultTxt = "none";
+
+  var snackBar = SnackBar(content: Text(resultTxt));
+  ScaffoldMessenger.of(context).showSnackBar(snackBar);
+}
+
+void chosenNodes(BuildContext context, TreeType tree) {
+  List<TreeType<EasyNodeType>> result = [];
+  var root = findRoot(tree);
+  returnChosenNodes(root, result);
+  String resultTxt = "";
+  for (var e in result) {
+    resultTxt += "${e.data.title}\n\n";
+  }
+  if (resultTxt.isEmpty) resultTxt = "none";
+
+  var snackBar = SnackBar(content: Text(resultTxt));
+  ScaffoldMessenger.of(context).showSnackBar(snackBar);
+}
+
+void searchLeaves(BuildContext context, String txt, TreeType tree) {
+  List<TreeType<EasyNodeType>> result = [];
+  var root = findRoot(tree);
+  searchLeavesWithTitleDFS(root, txt, result);
+  String resultTxt = "";
+  for (var e in result) {
+    resultTxt += "${e.data.title}\n\n";
+  }
+  if (resultTxt.isEmpty) resultTxt = "none";
+
+  var snackBar = SnackBar(content: Text(resultTxt));
+  ScaffoldMessenger.of(context).showSnackBar(snackBar);
+}
+
+void searchNodes(BuildContext context, String txt, TreeType tree) {
+  List<TreeType<EasyNodeType>> result = [];
+  var root = findRoot(tree);
+  searchAllTreesWithTitleDFS(root, txt, result);
+  String resultTxt = "";
+  for (var e in result) {
+    resultTxt += "${e.data.title}\n\n";
+  }
+  if (resultTxt.isEmpty) resultTxt = "none";
+
+  var snackBar = SnackBar(content: Text(resultTxt));
+  ScaffoldMessenger.of(context).showSnackBar(snackBar);
+}