+++ example/lib/screens/expandable_multiple_choice/expandable_multiple_choice_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../widgets/custom_outlined_btn.dart';
+import '../../widgets/btn1.dart';
 import 'ex_expandable_tree_expanded_screen.dart';
 import 'ex_expandable_tree_screen.dart';
 
@@ -24,12 +24,12 @@ class _ExpandableMultipleChoiceScreenState
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            CustomOutlinedButton(
+            Btn1(
               screen: ExExpandableTreeScreen(),
               title: "Expandable Tree\nmultiple choice - parse data 1 time",
             ),
             SizedBox(height: 20),
-            CustomOutlinedButton(
+            Btn1(
               screen: ExExpandableTreeExpandedScreen(),
               title:
                   "Expandable Tree expands first\nmultiple choice - parse data 1 time",