+++ example/lib/widgets/btn1.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
-class CustomOutlinedButton extends StatelessWidget {
+class Btn1 extends StatelessWidget {
   final String title;
   final StatefulWidget screen;
 
-  const CustomOutlinedButton({
+  const Btn1({
     super.key,
     required this.screen,
     required this.title,
@@ -13,12 +13,17 @@ class CustomOutlinedButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return OutlinedButton(
+      style: OutlinedButton.styleFrom(
+        backgroundColor: Colors.white,
+        padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
+      ),
       onPressed: () => Navigator.of(context).push(
         MaterialPageRoute(builder: (_) => screen),
       ),
       child: Text(
         title,
         textAlign: TextAlign.center,
+        style: const TextStyle(color: Colors.black, fontSize: 16),
       ),
     );
   }