+++ example/lib/widgets/btn3.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+class Btn3 extends StatelessWidget {
+  final String title;
+  final VoidCallback onTap;
+
+  const Btn3({
+    super.key,
+    required this.title,
+    required this.onTap,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return OutlinedButton(
+      style: OutlinedButton.styleFrom(
+        backgroundColor: Colors.blue.withOpacity(0.1),
+        padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
+      ),
+      onPressed: onTap,
+      child: FittedBox(
+        fit: BoxFit.scaleDown,
+        child: Text(
+          title,
+          textAlign: TextAlign.center,
+          style: TextStyle(
+            color: Colors.red[800],
+            fontSize: 16,
+            fontWeight: FontWeight.bold,
+          ),
+          maxLines: 1,
+        ),
+      ),
+    );
+  }
+}