+++ example/lib/widgets/divider.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+Widget get divider => const Divider(
+      thickness: 2,
+      height: 20,
+      color: Colors.black,
+    );