+++ lib/recursive_tree_flutter.dart
@@ -5,19 +5,14 @@
 
 library recursive_tree_flutter;
 
-export 'functions/tree_traversal_functions.dart';
-export 'functions/tree_update_functions.dart';
+export 'package:unit_tree_data_structure/unit_tree_data_structure.dart';
 
-export 'utils/tree_view_properties.dart';
+export 'utils/ui_properties.dart';
 
-export 'models/abstract_node_type.dart';
-export 'models/tree_type.dart';
 export 'models/typedef.dart';
 
-export 'views/vts/show_vts_department_tree.dart';
-export 'views/vts/vts_department_tree_widget.dart';
+export 'widgets/expanded_variant/expandable_tree_widget.dart';
+export 'widgets/expanded_variant/expandable_tree_mixin.dart';
 
-export 'views/expandable_tree_widget.dart';
-export 'views/lazy_stack_widget.dart';
-export 'views/expandable_tree_mixin.dart';
-export 'views/stack_widget.dart';
+export 'widgets/stack_variant/lazy_stack_widget.dart';
+export 'widgets/stack_variant/stack_widget.dart';